remake

+++ src/s21_atan.c
@@ -2,8 +2,8 @@
 
 #define ACCURACY 0.28
 
-double s21_atan(double x) {
-  double result, abs_x;
+long double s21_atan(double x) {
+  long double result, abs_x;
   abs_x = s21_fabs(x);
   if(abs_x < 1) {
     result = x / (1 + ACCURACY * x * x);
